Massive receive_frame refactor from Michael Fischer.
This deals with ActiveMQ sending 'buggy' STOMP frames:

Delving into the problem a bit more, I discovered that ActiveMQ was
issuing STOMP frames that were terminated not just with a NULL
character, but also a newline thereafter.  After the CONNECTED frame
was received, this was causing select() to indicate that the file
descriptor was ready, but before an entirely new frame had been sent.
Since Net::Stomp::Frame::_readline loops on sysread(), this would then
block.