Add support for 'nack' method to
negatively acknowledge a received frame.  Like the 'ack' method,
it should only be used if you requested client acks - the server may
barf if you start trying to ack/nack an auto-ack'd message.

Depending on the server, additional properties may be used, for example
RabbitMQ supports a 'requeue' value that can be used to cause the message
to either be dropped, or placed back on the queue.