Allow "failover" url to work with "ssl" and "ssl_options" parameters
The "ssl" and "ssl_options" parameters were ignored when connecting
using a "failover" URL. Now the SSL options are set for each host which
causes an SSL connection to be used, just like if an explicit "hosts"
list with SSL settings was used when creating a Net::Stomp object. This
also sort of fixes old CPAN RT #128661 in that it allows failover URL to
perform SSL connections.

Also added an extra check that we have a valid socket before calling
setsockopt SO_KEEPALIVE on it. This fixes old CPAN RT #118018.