Only update the _alive status when sending data
Brokers like ArtemisMQ seem to only care about clients sending data as a
sign the client is still alive, not that the broker forwarded tons of
messages to the client. In order for a primarily recieve-only client to
work properly with such a broker using server-side automatic acks that
client must send heartbeats. The fact that we were setting the _alive
time also when receiving messages prevented send_heartbeat from sending
the proper heartbeats since.